feat: surface missing PXI model credentials
