+++ app/src/components/generative/index.ts
@@ -1,3 +1,4 @@
 export * from "./ToolChoiceSelector";
 export * from "./GenerativeProviderIcon";
 export * from "./ModelMenu";
+export * from "./ProviderCredentialsDialog";
